pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "RAGtor"
version = "0.1.0"
description = "Initial Implementation of RAPTOR with Ollama and Open-Source Models."
authors = [{name = "Alberto Lorente Galé", email = "albertolorente97@gmail.com"}]
keywords = ["Raptor", "RAG"]
readme = "README.md"

requires-python="==3.12"
dependencies = [
"faiss-cpu==1.10.0",
"langchain-community==0.3.26",
"langchain-core==0.3.66 ",
"langchain-ollama==0.3.3",
"langchain-text-splitters==0.3.8",
"load==1.0.14",
"nltk==3.9.1",
"numpy==2.3.1",
"ollama==0.5.1",
"pdf2image==1.17.0",
"pdfplumber==0.11.7",
"Pillow==11.2.1",
"pydantic==2.11.7",
"scikit-learn==1.7.0",
"torch",
"wand==0.6.13",
"sentence-transformers",
"langchain-huggingface", 
"accelerate"]

[project.scripts]
ragtor = "ragtor.app:main"

[project.urls]
Repository = "https://github.com/alberto-lorente/RAGtor.git"

[tool.setuptools]
py-modules = ['ragtor']

